refactor: separate depedency map code
